description_generation_for_a_content/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

numpy
pandas

# 🔊 Whisper
openai-whisper

# 🎥 YouTube download
yt-dlp

# 🎬 Video & audio
moviepy
ffmpeg-python
imageio-ffmpeg
soundfile
librosa

# 🤖 NLP
transformers
sentencepiece

# 🔥 PyTorch
torch
torchaudio

# 🧩 utils
pydantic
tqdm